+++ backend/app-orchestrator-service/src/main/kotlin/services/ResourceStore.kt
@@ -213,7 +213,7 @@ class ResourceStore<T>(
     }
 
     suspend fun synchronizeNow() {
-        queries.withSession { session ->
+        queries.withSession("synchronizeNow") { session ->
             syncTable(session, uidShortcut)
             syncTable(session, pidShortcut)
         }
@@ -1129,7 +1129,7 @@ class ResourceStore<T>(
     private suspend fun initializeIdIndex(baseId: Long): IdIndex {
         val result = IdIndex(baseId)
 
-        queries.withSession { session ->
+        queries.withSession("initializeIdIndex") { session ->
             queries.loadIdIndex(session, type, baseId, baseId + IdIndex.BLOCK_SIZE) { id, ref, isUser ->
                 result.register(id, ref, isUser)
             }
@@ -1232,7 +1232,7 @@ class ResourceStore<T>(
 
             val result = ProviderIndex(provider)
 
-            queries.withSession { session ->
+            queries.withSession("findOrLoadProviderIndex") { session ->
                 queries.loadProviderIndex(session, type, provider) { uid, pid ->
                     result.registerUsage(uid, pid)
                 }
@@ -1361,7 +1361,7 @@ class ResourceStoreBucket<T>(
             }
         }
 
-        queries.withSession { session ->
+        queries.withSession("evictAll") { session ->
             var current: ResourceStoreBucket<T> = this
             while (true) {
                 current.synchronizeToDatabase(session, hasLock = true)
@@ -1406,7 +1406,7 @@ class ResourceStoreBucket<T>(
         mutex.withLock {
             if (evicted) throw EvictionException()
 
-            val nextId = queries.withSession { session ->
+            val nextId = queries.withSession("load(minimumId = $minimumId)") { session ->
                 queries.loadResources(session, this, minimumId)
             }
 
@@ -2253,7 +2253,7 @@ object ResourceIdAllocator {
     private suspend fun init(queries: ResourceStoreDatabaseQueries<*>) {
         initMutex.withLock {
             if (idAllocator.get() >= 0L) return
-            queries.withSession { session ->
+            queries.withSession("init") { session ->
                 idAllocator.set(max(5_000_000L, queries.loadMaximumId(session)))
             }
         }
@@ -2355,7 +2355,7 @@ fun Permission.toByte(): Byte {
 }
 
 interface ResourceStoreDatabaseQueries<T> {
-    suspend fun startTransaction(): Any
+    suspend fun startTransaction(reason: String): Any
     suspend fun commitTransaction(transaction: Any)
     suspend fun abortTransaction(transaction: Any)
     suspend fun loadResources(transaction: Any, bucket: ResourceStoreBucket<T>, minimumId: Long): Long?
@@ -2378,8 +2378,8 @@ interface ResourceStoreDatabaseQueries<T> {
     suspend fun loadMaximumId(transaction: Any): Long
 }
 
-suspend inline fun <R> ResourceStoreDatabaseQueries<*>.withSession(block: (Any) -> R): R {
-    val session = startTransaction()
+suspend inline fun <R> ResourceStoreDatabaseQueries<*>.withSession(reason: String = "unknown", block: (Any) -> R): R {
+    val session = startTransaction(reason)
     var success = true
     try {
         return block(session)
@@ -2431,7 +2431,7 @@ class ResourceStoreDatabaseQueriesImpl<T>(
         // NOTE(Dan): We do not need to check the table since it will fail the hasTrigger check if it contains an
         // invalid reference.
 
-        db.withSession { session ->
+        db.withSession(reason = "initializeEvictionTriggers") { session ->
             val hasTrigger = session.sendPreparedStatement(
                 { setParameter("table", qualifiedTable) },
                 """
@@ -2506,8 +2506,8 @@ class ResourceStoreDatabaseQueriesImpl<T>(
         }
     }
 
-    override suspend fun startTransaction(): Any {
-        val session = db.openSession()
+    override suspend fun startTransaction(reason: String): Any {
+        val session = db.openSession(reason = "ResourceStore." + reason)
         db.openTransaction(session)
         session.sendQuery("set local ucloud.performed_by_cache to 'true';")
         return session