+++ provider-integration/im2/pkg/im/config/format.go
@@ -8,13 +8,14 @@ import (
 	"encoding/pem"
 	"errors"
 	"fmt"
-	"gopkg.in/yaml.v3"
 	"net"
 	"os"
 	"path"
 	"path/filepath"
 	"regexp"
 	"strings"
+	"gopkg.in/yaml.v3"
+	fnd "ucloud.dk/pkg/foundation"
 	"ucloud.dk/pkg/log"
 	"ucloud.dk/pkg/util"
 )
@@ -364,8 +365,15 @@ func decode(filePath string, node *yaml.Node, result any, success *bool) {
 	}
 }
 
-func readAndParse(filePath string) *yaml.Node {
+func readAndParse(configDir string, name string) (string, *yaml.Node) {
+	filePath := filepath.Join(configDir, name+".yml")
 	fileBytes, err := os.ReadFile(filePath)
+
+	if err != nil {
+		filePath = filepath.Join(configDir, name+".yaml")
+		fileBytes, err = os.ReadFile(filePath)
+	}
+
 	if err != nil {
 		reportError(
 			filePath,
@@ -376,7 +384,7 @@ func readAndParse(filePath string) *yaml.Node {
 			os.Getgid(),
 			err.Error(),
 		)
-		return nil
+		return filePath, nil
 	}
 
 	var document yaml.Node
@@ -390,10 +398,10 @@ func readAndParse(filePath string) *yaml.Node {
 				"Underlying error message: %v.",
 			err.Error(),
 		)
-		return nil
+		return filePath, nil
 	}
 
-	return &document
+	return filePath, &document
 }
 
 func requireSecrets(reason string) (string, *yaml.Node) {
@@ -413,6 +421,11 @@ func Parse(serverMode ServerMode, configDir string) bool {
 	if Mode == ServerModeServer {
 		filePath := filepath.Join(configDir, "secrets.yml")
 		fileBytes, err := os.ReadFile(filePath)
+
+		if err != nil {
+			filePath = filepath.Join(configDir, "secrets.yaml")
+			fileBytes, err = os.ReadFile(filePath)
+		}
 
 		if err == nil {
 			var document yaml.Node
@@ -435,8 +448,7 @@ func Parse(serverMode ServerMode, configDir string) bool {
 	}
 
 	if Mode == ServerModeServer {
-		filePath := filepath.Join(configDir, "server.yml")
-		document := readAndParse(filePath)
+		filePath, document := readAndParse(configDir, "server")
 		if document == nil {
 			return false
 		}
@@ -448,8 +460,7 @@ func Parse(serverMode ServerMode, configDir string) bool {
 		Server = &server
 	}
 
-	filePath := filepath.Join(configDir, "config.yml")
-	document := readAndParse(filePath)
+	filePath, document := readAndParse(configDir, "config")
 	if document == nil {
 		return false
 	}
@@ -691,6 +702,10 @@ type ProviderConfiguration struct {
 			RetentionPeriodInDays int  `yaml:"retentionPeriodInDays"`
 		}
 	}
+
+	Maintenance struct {
+		UserAllowList []string
+	}
 }
 
 func parseProvider(filePath string, provider *yaml.Node) (bool, ProviderConfiguration) {
@@ -767,6 +782,19 @@ func parseProvider(filePath string, provider *yaml.Node) (bool, ProviderConfigur
 					return false, cfg
 				}
 			}
+		}
+	}
+
+	{
+		// Maintenance section
+		maintenance, _ := getChildOrNil(filePath, provider, "maintenance")
+		if maintenance != nil {
+			allowListNode := requireChild(filePath, maintenance, "userAllowList", &success)
+
+			var userAllowList []string
+			decode(filePath, allowListNode, &userAllowList, &success)
+
+			cfg.Maintenance.UserAllowList = userAllowList
 		}
 	}
 
@@ -987,12 +1015,13 @@ type IdentityManagementScripted struct {
 }
 
 type IdentityManagementFreeIPA struct {
-	Url        string
-	VerifyTls  bool
-	CaCertFile util.Option[string]
-	Username   string
-	Password   string
-	GroupName  string
+	Url             string
+	VerifyTls       bool
+	CaCertFile      util.Option[string]
+	Username        string
+	Password        string
+	GroupName       string
+	ProjectStrategy fnd.ProjectTitleStrategy
 }
 
 func (m *IdentityManagement) Scripted() *IdentityManagementScripted {
@@ -1082,6 +1111,25 @@ func parseIdentityManagementFreeIpa(filePath string, node *yaml.Node) (bool, Ide
 		result.GroupName = groupName
 	} else {
 		result.GroupName = "ucloud_users"
+	}
+
+	titleStrategy := optionalChildText(filePath, node, "projectStrategy", &success)
+	if titleStrategy != "" {
+		switch titleStrategy {
+		case "Default":
+			result.ProjectStrategy = fnd.ProjectTitleDefault
+
+		case "Date":
+			result.ProjectStrategy = fnd.ProjectTitleDate
+
+		case "UUID":
+			result.ProjectStrategy = fnd.ProjectTitleUuid
+
+		default:
+			success = false
+			badNode, _ := getChildOrNil(filePath, node, "projectStrategy")
+			reportError(filePath, badNode, titleStrategy, "Unknown title strategy, use one of: 'Default', 'Date', 'UUID'")
+		}
 	}
 
 	return success, result